# api/swagger.yml
...
consumes:
- application/json
produces:
- application/json
swagger: "2.0"
info:
  description: API documentation
  title: github.com/kashguard/go-kms
  version: 0.1.0
paths:
...
  /api/v1/kms/keys/{keyId}:
    get:
      security:
      - Bearer: []
      description: Get key metadata
      tags:
      - kms
      summary: Get key
      operationId: GetKeyRoute
      parameters:
      - type: string
        description: Key ID
        name: keyId
        in: path
        required: true
      responses:
        "200":
          description: GetKeyResponse
          schema:
            $ref: '#/definitions/getKeyResponse'
        "401":
          description: PublicHTTPError
          schema:
            $ref: '#/definitions/publicHttpError'
        "404":
          description: Key not found
          schema:
            $ref: '#/definitions/publicHttpError'
    put:
      security:
      - Bearer: []
      description: Update key metadata
      tags:
      - kms
      summary: Update key
      operationId: PutUpdateKeyRoute
      parameters:
      - type: string
        description: Key ID
        name: keyId
        in: path
        required: true
      - name: Payload
        in: body
        required: true
        schema:
          $ref: '#/definitions/putUpdateKeyPayload'
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/getKeyResponse'
        "401":
          description: PublicHTTPError
          schema:
            $ref: '#/definitions/publicHttpError'
        "403":
          description: Policy denied
          schema:
            $ref: '#/definitions/publicHttpError'
        "404":
          description: Key not found
          schema:
            $ref: '#/definitions/publicHttpError'
    delete:
      security:
      - Bearer: []
      description: Delete key (schedule for deletion)
      tags:
      - kms
      summary: Delete key
      operationId: DeleteKeyRoute
      parameters:
      - type: string
        description: Key ID
        name: keyId
        in: path
        required: true
      responses:
        "204":
          description: Success
        "401":
          description: PublicHTTPError
          schema:
            $ref: '#/definitions/publicHttpError'
        "403":
          description: Policy denied
          schema:
            $ref: '#/definitions/publicHttpError'
        "404":
          description: Key not found
          schema:
            $ref: '#/definitions/publicHttpError'
...
  /api/v1/kms/secrets:
    post:
      security:
      - Bearer: []
      description: Create a new secret and store encrypted data
      tags:
      - kms
      summary: Create secret
      operationId: PostCreateSecretRoute
      parameters:
      - name: Payload
        in: body
        required: true
        schema:
          $ref: '#/definitions/postCreateSecretPayload'
      responses:
        "201":
          description: CreateSecretResponse
          schema:
            $ref: '#/definitions/createSecretResponse'
        "400":
          description: PublicHTTPValidationError
          schema:
            $ref: '#/definitions/publicHttpValidationError'
        "401":
          description: PublicHTTPError
          schema:
            $ref: '#/definitions/publicHttpError'
        "409":
          description: Secret already exists
          schema:
            $ref: '#/definitions/publicHttpError'
  /api/v1/kms/secrets/{keyId}:
    get:
      security:
      - Bearer: []
      description: Get secret and decrypt data
      tags:
      - kms
      summary: Get secret
      operationId: GetSecretRoute
      parameters:
      - type: string
        description: Secret key ID
        name: keyId
        in: path
        required: true
      responses:
        "200":
          description: GetSecretResponse
          schema:
            $ref: '#/definitions/getSecretResponse'
        "401":
          description: PublicHTTPError
          schema:
            $ref: '#/definitions/publicHttpError'
        "404":
          description: Secret not found
          schema:
            $ref: '#/definitions/publicHttpError'
    put:
      security:
      - Bearer: []
      description: Update secret data
      tags:
      - kms
      summary: Update secret
      operationId: PutUpdateSecretRoute
      parameters:
      - type: string
        description: Secret key ID
        name: keyId
        in: path
        required: true
      - name: Payload
        in: body
        required: true
        schema:
          $ref: '#/definitions/putUpdateSecretPayload'
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/getSecretResponse'
        "401":
          description: PublicHTTPError
          schema:
            $ref: '#/definitions/publicHttpError'
        "404":
          description: Secret not found
          schema:
            $ref: '#/definitions/publicHttpError'
    delete:
      security:
      - Bearer: []
      description: Delete secret
      tags:
      - kms
      summary: Delete secret
      operationId: DeleteSecretRoute
      parameters:
      - type: string
        description: Secret key ID
        name: keyId
        in: path
        required: true
      responses:
        "204":
          description: Success
        "401":
          description: PublicHTTPError
          schema:
            $ref: '#/definitions/publicHttpError'
        "404":
          description: Secret not found
          schema:
            $ref: '#/definitions/publicHttpError'
    head:
      security:
      - Bearer: []
      description: Check if secret exists
      tags:
      - kms
      summary: Check secret exists
      operationId: HeadSecretExistsRoute
      parameters:
      - type: string
        description: Secret key ID
        name: keyId
        in: path
        required: true
      responses:
        "200":
          description: Secret exists
        "401":
          description: PublicHTTPError
          schema:
            $ref: '#/definitions/publicHttpError'
        "404":
          description: Secret not found
...
definitions:
...
  createSecretResponse:
    type: object
    required:
    - key_id
    properties:
      created_at:
        type: string
        format: date-time
        example: "2025-11-26T10:00:00Z"
      key_id:
        type: string
        example: wallet-keystore-00000000-0000-0000-0000-000000000001
...
  getKeyResponse:
    type: object
    required:
    - key_id
    - key_type
    - key_state
    properties:
      alias:
        type: string
      created_at:
        type: string
        format: date-time
      description:
        type: string
      key_id:
        type: string
        example: key-1234567890abcdef
      key_state:
        type: string
        example: Enabled
      key_type:
        type: string
        example: AES_256
      latest_version:
        type: integer
        example: 1
      tags:
        type: object
        additionalProperties:
          type: string
      updated_at:
        type: string
        format: date-time
...
  getSecretResponse:
    type: object
    required:
    - key_id
    - data
    properties:
      data:
        description: 解密后的数据（base64编码）
        type: string
        format: byte
        example: eyJ2ZXJzaW9uIjozLCJpZCI6InV1aWQiLCJjcnlwdG8iOnsiY2lwaGVydGV4dCI6ImVuY3J5cHRlZF9tbmVtb25pY19oZXgiLCJjaXBoZXJwYXJhbXMiOnsiaXYiOiJpdl9oZXgifSwiY2lwaGVyIjoiYWVzLTEyOC1jdHIiLCJrZGYiOiJzY3J5cHQiLCJrZGZwYXJhbXMiOnsiZGtsZW4iOjMyLCJzYWx0Ijoic2FsdF9oZXgiLCJuIjoyNjIxNDQsInIiOjgsInAiOjF9LCJtYWMiOiJtYWNfaGV4In19fQ==
      key_id:
        type: string
        example: wallet-keystore-00000000-0000-0000-0000-000000000001
      updated_at:
        type: string
        format: date-time
        example: "2025-11-26T10:00:00Z"
...
  httpValidationErrorDetail:
    type: object
    required:
    - key
    - in
    - error
    properties:
      error:
        description: Error describing field validation failure
        type: string
      in:
        description: Indicates how the invalid field was provided
        type: string
      key:
        description: Key of field failing validation
        type: string
...
  postCreateSecretPayload:
    type: object
    required:
    - key_id
    - data
    properties:
      data:
        description: 要存储的数据（base64编码）
        type: string
        format: byte
        example: eyJ2ZXJzaW9uIjozLCJpZCI6InV1aWQiLCJjcnlwdG8iOnsiY2lwaGVydGV4dCI6ImVuY3J5cHRlZF9tbmVtb25pY19oZXgiLCJjaXBoZXJwYXJhbXMiOnsiaXYiOiJpdl9oZXgifSwiY2lwaGVyIjoiYWVzLTEyOC1jdHIiLCJrZGYiOiJzY3J5cHQiLCJrZGZwYXJhbXMiOnsiZGtsZW4iOjMyLCJzYWx0Ijoic2FsdF9oZXgiLCJuIjoyNjIxNDQsInIiOjgsInAiOjF9LCJtYWMiOiJtYWNfaGV4In19fQ==
      key_id:
        description: Secret 标识符
        type: string
        example: wallet-keystore-00000000-0000-0000-0000-000000000001
...
  publicHttpError:
    type: object
    required:
    - status
    - type
    - title
    properties:
      detail:
        description: More detailed, human-readable, optional explanation of the error
        type: string
        example: User is lacking permission to access this resource
      status:
        description: HTTP status code returned for the error
        type: integer
        format: int64
        maximum: 599
        minimum: 100
        x-go-name: Code
        example: 403
      title:
        description: Short, human-readable description of the error
        type: string
        example: Forbidden
      type:
        $ref: '#/definitions/publicHttpErrorType'
  publicHttpErrorType:
    description: Type of error returned, should be used for client-side error handling
    type: string
    enum:
    - generic
    - PUSH_TOKEN_ALREADY_EXISTS
    - OLD_PUSH_TOKEN_NOT_FOUND
    - ZERO_FILE_SIZE
    - USER_DEACTIVATED
    - INVALID_PASSWORD
    - NOT_LOCAL_USER
    - TOKEN_NOT_FOUND
    - TOKEN_EXPIRED
    - USER_ALREADY_EXISTS
    - MALFORMED_TOKEN
    - LAST_AUTHENTICATED_AT_EXCEEDED
    - MISSING_SCOPES
  publicHttpValidationError:
    type: object
    required:
    - validationErrors
    allOf:
    - $ref: '#/definitions/publicHttpError'
    properties:
      validationErrors:
        description: List of errors received while validating payload against schema
        type: array
        items:
          $ref: '#/definitions/httpValidationErrorDetail'
  putUpdateKeyPayload:
    type: object
    properties:
      description:
        description: 密钥描述
        type: string
      policy_id:
        description: 策略ID
        type: string
      tags:
        type: object
        additionalProperties:
          type: string
...
  putUpdateSecretPayload:
    type: object
    required:
    - data
    properties:
      data:
        description: 要更新的数据（base64编码）
        type: string
        format: byte
        example: eyJ2ZXJzaW9uIjozLCJpZCI6InV1aWQiLCJjcnlwdG8iOnsiY2lwaGVydGV4dCI6ImVuY3J5cHRlZF9tbmVtb25pY19oZXgiLCJjaXBoZXJwYXJhbXMiOnsiaXYiOiJpdl9oZXgifSwiY2lwaGVyIjoiYWVzLTEyOC1jdHIiLCJrZGYiOiJzY3J5cHQiLCJrZGZwYXJhbXMiOnsiZGtsZW4iOjMyLCJzYWx0Ijoic2FsdF9oZXgiLCJuIjoyNjIxNDQsInIiOjgsInAiOjF9LCJtYWMiOiJtYWNfaGV4In19fQ==
...
securityDefinitions:
  Bearer:
    description: |-
      Access token for application access, **must** include "Bearer " prefix.
      Example: `Bearer b4a94a42-3ea2-4af3-9699-8bcbfee6e6d2`
    type: apiKey
    name: Authorization
    in: header
    x-keyPrefix: 'Bearer '
  Management:
    description: Management secret, used for monitoring and infrastructure related
      calls
    type: apiKey
    name: mgmt-secret
    in: query